+++ mongock-spring/src/main/java/com/github/cloudyrock/mongock/decorator/MongoDbFactoryDecorator.java
@@ -1,8 +1,11 @@
 package com.github.cloudyrock.mongock.decorator;
 
 import com.github.cloudyrock.mongock.decorator.impl.MongoDataBaseDecoratorImpl;
+import com.github.cloudyrock.mongock.decorator.impl.MongoDbFactoryDecoratorImpl;
 import com.github.cloudyrock.mongock.decorator.util.MongockDecoratorBase;
+import com.mongodb.ClientSessionOptions;
 import com.mongodb.DB;
+import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoDatabase;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.support.PersistenceExceptionTranslator;
@@ -27,6 +30,17 @@ public interface MongoDbFactoryDecorator extends MongockDecoratorBase<MongoDbFac
 
   @Override
   default DB getLegacyDb() {
-    throw new UnsupportedOperationException("Removed DB support from Mongock due to depregated API. Please use MongoDatabase instead");
+    throw new UnsupportedOperationException("Removed DB support from Mongock due to deprecated API. Please use MongoDatabase instead");
+  }
+
+
+  @Override
+  default ClientSession getSession(ClientSessionOptions clientSessionOptions) {
+    return getInvoker().invoke(()-> getImpl().getSession(clientSessionOptions));
+  }
+
+  @Override
+  default MongoDbFactory withSession(ClientSession clientSession) {
+    return new MongoDbFactoryDecoratorImpl(getImpl().withSession(clientSession), getInvoker());
   }
 }