+++ mongock-spring/src/main/java/com/github/cloudyrock/mongock/decorator/impl/MongoDbFactoryDecoratorImpl.java
@@ -2,6 +2,8 @@ package com.github.cloudyrock.mongock.decorator.impl;
 
 import com.github.cloudyrock.mongock.decorator.MongoDbFactoryDecorator;
 import com.github.cloudyrock.mongock.decorator.util.MethodInvoker;
+import com.mongodb.ClientSessionOptions;
+import com.mongodb.client.ClientSession;
 import org.springframework.data.mongodb.MongoDbFactory;
 
 public class MongoDbFactoryDecoratorImpl implements MongoDbFactoryDecorator {
@@ -23,4 +25,5 @@ public class MongoDbFactoryDecoratorImpl implements MongoDbFactoryDecorator {
   public MethodInvoker getInvoker() {
     return invoker;
   }
+
 }